Add a with-fixtures helper for running tests in fixtures
Unfortunately, fixture execution is normally tied to `deftest`. Because
`deftest` can't be given a dynamic variable name, it can't really be
used in a loop to test multiple pieces of code that do the same thing.
`with-fixtures` is a new macro which separates invoking fixtures from
`deftest`. Any code inside a call to `with-fixtures` will be surrounded
by the fixtures for the current namespace.