Add a v2 of 'replace catalog' command that doesn't accept classes/tags
This version is also stricter about which fields are present, failing if
there are any keys missing or any extra keys supplied. There is also a
refactor of catalog parsing included, which replaces
`parse-from-json-string` and `parse-from-json-obj` with a multimethod
`parse-catalog` that can handle parsing from a string, or from version 1
and version 2 wire formats. The result will be a transformed and
validated catalog in version 2 of the PuppetDB format, regardless of the
input.

The 'replace catalog' v2 handler is identical to the v1 handler, and the
actual work has been extracted to `replace-catalog*` (and probably needs
somewhere better to live in the future). The only difference between the
two command processors is the version of input they take, though they
actually both handle it the same, by passing it to `parse-catalog`,
which always results in the same format.