Allow command payloads to be embedded objects
Previously these were required to be JSON stringified, and would be
automatically serialized by our command helpers. Now, the serialization
must be done by the producer if it's wanted, and the command is free to
choose whether its payload ought to be an object or the string
representation of that object.

For correct backward compatibility, 'replace catalog' v1 now explicitly
requires that its payload be a string. The v2 of the command allows
either a string *or* an object. Other commands are unchanged.