Add UNIQUE constraint to resource_events table
We had a decent amount of discussion around the fact that the
"resource_events" table, in its current form, can't have a
meaningful primary key.  The "correct" primary key would be
(report_id + resource_type + resource_title + property), but
property is nullable because it is null for skipped resources.
We discussed the idea of splitting this table into multiple
tables, but ultimately decided that for now, a UNIQUE constraint
would be good enough.