Don't ignore messages that are processed after deactivation
Prior to this commit, if a facts/catalog command was sent
before a deactivate command but processed *after* the
deactivate command, it would be ignored.  This commit
changes the behavior so that the command will still be
processed (without reactivating the node), and refactors
a small bit of repetetive code.