Get rid of `description` field in reports table
We'd added a `description` field to the table back
when it was still called `event_groups` and we were
trying to be generic.  It seemed like it might be
useful for different event sources to have a place
to provide some human-readable explanation of what
triggered the events.

Now that the table is more specific (`reports`),
this column no longer feels useful, so I'm getting
rid of it.  We can always add something like it
back in later when we start supporting other
event sources.