Clean up query parameters for `reports` query endpoint
This commit makes the input to the report query a bit more
similar to our other existing queries; prior to this commit
I had a (not-yet-implemented) `query` parameter, and a separate
parameter called `report-id`.  This was different from our
existing endpoints which generally all just except one parameter
called `query` and then expect the expression therein to be used
to filter.  Now the reports endpoint works this way as well.

In addition, I've gotten rid of the concept of querying for
report summaries by `report-id`, which seems useless.  In its
place I've added support for querying by `certname`, which
seems much more useful for UI tools.

Still need to do a similar refactor / cleanup for the `events`
endpoint.