Refactor query implementation
This reduces duplication by creating a com.puppetlabs.puppetdb.query
namespace, which now holds the actual query operators. The operators are
now explicitly listed per-endpoint, rather than using a multimethod.
This allows us to separate v1 from v2 query, so that we can maintain
backward compatibility by not supporting subqueries for resources in the
v1 API. It also means we only have a single implementation of 'and',
'or', and 'not', as well as the general query framework.

Query is now driven by the query/resource-query->sql and
query/fact-query->sql functions, which accept as an argument the map of
operators to function implementations to use.

This commit also changes the subquery API itself a bit. The column to
correlate with for 'in-result' and 'project' no longer needs to be
namespaced, and 'node' is no longer supported in place of 'certname'
(for consistency).