(#17397) DB pool setup fails with numeric username/pw
Our config file parsing code tries to be clever and automatically
convert things that look like numbers into proper numbers. That's never
been a problem. However, interop with Java requires specific types
(nothing is auto-coerced or converted).

This bug happens during `make-connection-pool`. If the username or
password is numeric, when parsing the configuration file they're turned
into numbers. Then when we interop with BoneCP, we get an error because
we're passing in numbers when strings are expected.

As these are the only 2 possible properties that may be auto-converted
to numbers, simply coercing them to the correct type is sufficient.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>