Fix JDBC connection pool test on Postgres
When using Postgres as the test database, there's no guarantee the user this
test is trying to connect as exists. If it doesn't, the test hangs, trying
perpetually to connect. Given that the only thing we care about is that we can
build the appropriate connection pool, there's no need to actually connect. So
now we just build the object and go on, so this will work on both test
databases.