Swap order of keys/map arguments in mapvals signature
Looking at how the arguments to functions like `reduce` are ordered,
we realized that in order to support threading operations the
map should either always be the first or the last argument for
functions like mapvals.  The existing implementation had one
signature where the map was the second of three arguments, which
didn't make sense... this just swaps them.