Remove `id` from report model
Because I was including the `id` field in the model for reports,
I was having to pass it around to a bunch of functions that
shouldn't have needed it in order for reports to be validated
successfully.  Pulling the field out of the model allows those
functions to be a bit more straightforward, and we can simply
accept the `id` as an extra parameter to the `add-report!`
function.