Get rid of parse-from-json-string
Since we've recently started moving away from the idea of forcing
command payloads to be serialized to a String, this commit completely
removes the support for that for report storage.  The payload is
simply a map now.

Also includes a few other pre/post conditions and tidying up as
suggested in comments on the pull request.