Use a hash of the report contents as the report id
This commit adds functions for generating a sha1 hash based on the
contents of a report, and uses these functions to generate the
value that we will use to populate the report's `id` field, rather
than simply using the command's id (uuid).  This will allow us
to detect attempts to submit duplicate reports, and generally
have a better means of determining the uniqueness of an individual
report.

We may want to rename the field in the database schema from `id`
to something more descriptive of what we're actually storing there
now... perhaps `report_hash`.