(#17356) Queries should involve only active nodes
Previously, queries had to specify that they wanted to exclude
deactivated nodes. Unfortunately, it turned out that there are basically
zero use cases for wanting to query resources or facts of deactivated
nodes, and the overwhelmingly common use cases required an extra clause
to ensure that inactive nodes were excluded.

It's much nicer to just fiat that resource and fact queries only ever
touch active nodes. Common queries are now much shorter, and we
eliminate a source of user confusion.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>