Explicitly fail when trying to make a subquery against /v1/nodes
This lets the user know that their query isn't necessarily invalid, but
is just being run against the wrong version of the API.

Also quoted the operator in a few error messages ("Operator in is not
supported" is confusing) and fixed a keyword vs. string issue from a
mass find/replace.