Add HTTP endpoint for /facts/:name/:value
This will let you collect all the facts across a population that have a
particular name and value, like "give me all the facts you know about
with a name of 'operatingsystem' and a value of 'Debian'". It's an easy
way to find all the hosts with that fact, without having to do a big
query.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>