Fix json parsing for auto-deactivation of nodes during db gc
The command submission for "deactivate node" in the database
garbage collection loop was not json-serializing the node
name quite as many times as all of the other uses of that
command, so it was failing.

This commit adds in the extra JSON serialization and should
get things working again.  We might at some point want to
remove all of these seemingly redundant serializations
from all of the uses of this command, but this seemed like
the lower-impact change for the time being.