Reorganize report/event test namespaces
This commit creates some namespaces inside of `testutils` to hold
test utility functions for reports and events.  This allows us to
get rid of some duplicated utility functions and to put the actual
tests in the right places; prior to this there were a few tests that
should have been in the `storage` namespace that were not, simply
because they relied on some utility functions that they couldn't
easily access.