Add command-specific spool config settings
This commit only modifies the configuration settings code.  It
allows the definition of new `[facts]`, `[catalogs]`, and
`[reports]` settings in the `puppetdb.conf` file.  Each of these
sections supports a boolean setting named `spool`; these settings
will be used to control whether or not PuppetDB commands are spooled
to disk before submission and retried after failures, but this
commit does not implement that functionality yet.

The default for the `spool` setting is `true` for reports, and
`false` for catalogs and facts.