Add a maximum number of failures to the flush_commands method
In cases where PuppetDB has been down for an extended period of
time, you could end up with a large number of commands queued for
delivery.  Without some kind of maximum retries limitation, then
as long as PuppetDB remains down, every agent run that comes into
the master would cause PuppetDB to retry every single command that
was still in the queue.  Depending on how quickly the attempt to
connect to PuppetDB times out, this could cause huge performance
issues on the puppet master.

This commit adds a failure limit (currently hard-coded to 3) to
the command flushing logic.  Thus, if we try unsuccessfully to
flush 3 queued commands to PuppetDB, we'll stop trying until
the next time a flush is triggered.

Eventually we may want something even a bit more sophisticated
than this; this implementation will still cause 3 attempts on
every single agent run for the duration of the PuppetDB downtime.
We could build on this implementation by capturing a timestamp of
the last time that we encountered the maximum number of failures,
and don't retry again for a certain amount of time (5 minutes?)
after that.