Add "applied-migrations" function
This commit adds a function 'applied-migrations', which returns
a list of all of the migration versions that have been applied
to the existing database.  The idea is that we will use this
to replace the "schema-version" function, so that migrations
will run *all* missing migrations in order, rather than just
checking for the latest numeric value.