Change `submit_command` to an optimistic queueing strategy
Prior to this commit, the `submit_command` method would always
enqueue any queueable command immediately upon receipt--before
attempting to send it to puppetdb.

This commit changes the behavior to an optimistic approach, where
we assume the command will be submitted to PuppetDB successfully
and therefore we only enqueue the command if it fails.