HTTP /nodes endpoint should return node status objects
Previously, /v2/nodes would return a flat list of nodes. This patchset
changes it to instead return an array of status objects (like returned
in /v1/status).

Because this unifies the notion of a "status" object and a "node"
object, we can now just ditch the dedicated `query.status` namespace as
its functionality is now subsumed in `query.node`.

We can also get rid of the `/v2/status` endpoint. It's functionality is
duplicated by `/v2/nodes/<node name>`.