Change command queueing to ensure FIFO
This commit makes a slight tweak to the command queueing logic that
ensures we get first-in, first-out behavior.  (We basically accomplish
this by prefixing a timestamp at the beginning of the spool file
names, and then making sure that the iterator method sorts the
spooled files by their names before loading them.  I initially tried
to just use the file mtime but there was not enough precision.)