Refactor to support 'puppetdb retry' face/action
This commit moves the command submission logic into the Command
class, so that it is now entirely self-contained and no longer
has any dependencies on the indirector.  It also introduces a
Face called `puppetdb`, with a single action, `retry`, which
can be used to retry any failed PuppetDB commands that are
queued.