Add support for reports to the benchmark command
A report directory can now be specified with -R. If it is supplied
"store report" commands will be submitted every tick, alongside a
catalog submission. For consistency, this also changes the argument for
specifying a catalog directory to -C from -d.

If either reports or catalogs are not supplied, that kind of data will
be skipped, and only the other will be benchmarked.

Also improved error handling during submission, as they're done in
futures and thus would swallow all errors. Similarly, if an agent
experiences a failure, that will now be logged and the agent restarted.