Fix type for configuration-version
The configuration-version field on reports doesn't really have
any type requirements; any string should work.  There was a bug
where it was configured in the `Report` model to be of type
`datetime`; this appears to have been working so far because
all of the strings I've passed into it have been integers, which
the clojure time library must have been interpreting as
seconds-since-epoch or something, and thus they passed the
`datetime?` check.

This commit fixes it so that the model only validates that the
input is a `string`.  Also changes the examples to use a string
that would have failed the `datetime?` test, so that the tests
actually show this issue.