[packaging] Add cow update prior to pe deb build in Rakefile
Currently building debs for PE often fails the first time, because the cows in
use were previously used for FOSS products, and do not have the PE freight
repository enabled for dependency resolution. This has always been handled by
manually updating the cows, and then re-running the task. This begs the
question - why don't we automate that? This commit does so. It also allows the
arch on the cow build to be overridden with an environment variable, and
removes ENV['DIST'], which isn't really necessary since we explicitly set the
variable in the update call.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>