[packaging] Clear up @pe_version variable for PE deb building
We want to be able to override the default PE version we're building against
with an environment variable, and the current setup does not allow this (the
environment variable is always set to the hardcoded version). With this commit,
we use the ENV['PE_VER'] when available, otherwise default to 2.7. PE 2.7 is
the first version with wheezy PE packages built, thus it is the first one that
will be able to satisfy PE dependencies.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>