Clean up comments re: indirector voodoo
A while back, the utility module defined in `utils/puppetdb.rb`
was pretty tightly coupled with the implementation details of the
Puppet indirector class heirarchy, and the module could (mostly)
only be used as a mix-in on an indirector terminus.  Since then
it's been made to be a bit more general purpose so that it could
be used in other places as well (e.g. report processor).  However,
some of the comments in the code that were meant to explain the
coupling with the indirector were not updated when those changes
were made.  This commit simply removes the old obsolete comments.