Add utils for capturing log output during testing
This commit adds a namespace for logging-related test utility
functions.  Therein, it addsmacro and a couple of functions.
The macro, `with-log-output`, simplifies the process
of temporarily redirecting and capturing logging output during
tests.  The utility function `logs-matching`
provides a means for searching through the captured logs for log
entries whose log messages match a specified regular expression
pattern.