[packaging] Rewrite Rakefile for packaging repo, split out separate tasks
This commit rewrites the top-level Rakefile in puppetdb to use the packaging
repo. All debian/redhat methods and tasks supplied by the packaging repo are
removed. Because of the fairly specialized logic in the tarball composition
phase of puppetdb packaging, we take an approach similar to the one used in PE
live-managment packaging, which is to selectively undefine tasks loaded from
the packaging repo that we would rather override with ones specialized for this
project, e.g. everything up to tarball creation, and clean/clobber. This commit
also moves the remaining tasks out into their own files under tasks/ for
readability/organization and to make the Rakefile a little more sane.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>