[packaging] Fixup dependencies for pe-java, sles
This commit is essentially the work done by @haus in commits 90d5606 and
a29854c, with some minor modifications to redhat dependencies for sles. This
commit updates pe-puppetdb to depend on pe-java as opposed to system java, and
updates the sles packaging to name the appropriate packages providing the tools
we depend on.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>