[packaging] redhat spec fixups to remove erb os-based conditionals
The problem with evaluating os-based conditionals in the erb templating
is that it forces us to use different srpms for different OS's etc. This
becomes less sustainable as we try to automate more of this with the
packaging repo, which expects the specfile to take care of itself
regarding such things. This commit replaces some of the conditionals
with actual redhat spec conditionals. The biggest change is that of
changing the puppet_libdir assignment when building for pe to the
ruby sitelibdir, instead of a conditional on ruby version. Additionally,
on the redhat side, we support ruby 1.9 for fedora >= 17, so this commit
changes that conditional from an os_family fact check to a spec file
conditional, to avoid having to build different FOSS rpms for different
ruby versions.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>