[packaging] symlink pe-puppetdb redhat specfile to puppetdb
The packaging repo expects a specfile named after the project. A
dual-purpose one like puppetdb's isn't supported yet, so we need to one,
even it is just a symlink. However, there's a drawback. Ruby 1.9.3 won't
let us copy the symlink before the target and preserve attributes when
the target of the symlink doesn't exist yet, since it tries to set the
symlink metadata using a file that doesn't exist yet. This means we have
to use the system cp to create the dangling symlink in the workdir.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>