[packaging] Rework acceptance package build for packaging repo
This commit replaces much of the shell work to build packages with packaging
repo rake task invocations (though still via the acceptance package build shell
scripts). The packaging repo can do ssh remote building for us, so we don't
have to ssh out quite as much as we used to. One significant change, which we
may decide to undo later, is that the acceptance package builds for debian now
build puppetdb in every cow we ship.  This is technically the appropriate way
to accomplish this, as opposed to the single squeeze-based build we currently
ship for FOSS puppetdb to all dists. However, the downside is a dramatically
increased time to build all packages, with the current infrastructure setup.
However, the improvement in build times is actually a current project for
release, so the pain may be shortlived.