Rename `documentation` to `documentation.old`
This commit is in preparation for syncing our docs up with the ones
in the `puppet-docs` repo.  Just to make the git history a bit
easier to follow, I've renamed the docs directory to `documentation.old`;
the next commit will add the new docs from the other repo, and then
I'll delete the `old` directory in a final commit.