Copy latest docs from `puppet-docs` repo
We'd like for the docs in the puppetdb repo to always be up-to-date
with the latest code.  However, we also need to keep things in sync
with the documentation team and have a relatively sane way of
publishing the docs to the website.  We'd gotten pretty far out
of sync, so we just went through a process of merging all of the
latest diffs from our docs into the `puppet-docs` repo and reorganizing
things a bit over there.  This commit just copies in the result of that
work into our repo.

Going forward, all docs changes will be made here, and then when we're ready
to publish docs for a new release we'll just copy the whole docs folder
from here to a new, versioned subfolder in the `puppet-docs` repo.