Remove old documentation from repo
Now that we've synced up with the docs in the `puppet-docs` repo,
that version of our docs is the one we'll be maintaining going
forward.  This commit just gets rid of the old version.

Doing this in an isolated commit in hopes of making the git history
a bit easier to follow.