Transform README into more of a table-of-contents
The README had gotten extremely long, and in the docs repo
it had been broken up into several discrete documents.  Now
that we've synced down the docs from that repo, our README
was redundant.  This commit replaces it with basically a
table of contents with links to the rest of the documentation.

This commit also removes CHANGELOG.md, which was duplicated by
`documentation/release_notes.markdown`, and FAQ.md, which is
replaced by `documentation/puppetdb-faq.markdown`.