[packaging] Don't attempt to use params file inside build root
When ENV['PARAMS_FILE'] has a value, as it does with remote builds, the next
iteration of the packaging repo will attempt to read packaging data from the
path supplied. This environment variable remains set inside the buildroot, but
the path is no longer accessible/correct inside the buildroot. Luckily, since
we don't use anything other than defaults and the env vars passed in the rake
calls inside the buildroots, we don't need PARAMS_FILE anyway. This commit
ensures it is not set for this case.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>