(#18879) Remove the environment key from v1 catalogs
In some versions of Puppet, this key may be present. This fixes an issue
where older versions of the terminus (or otherwise perfectly legal v1
catalogs) would submit a catalog with this unexpected, but still not
allowed key. Now we explicitly remove it, so that when we call the v2
parse-catalog function, the catalog is legal.