Remove invalid handling of HTTP 404
Prior to this commit, we were catching HTTP 404 Not Found responses
and returning `nil`.  This was masking the fact that the URL for the
facts/find endpoint was invalid; also, the server actually just returns
a 200 with an empty array as the payload if you issue a well-formed
query for a node that puppetdb doesn't know about.