Update docs to include info about setting dashboard host via module
The current documentation does not reflect the fact that the module
provides a parameter for setting the listen address for the PuppetDB
dashboard.  It mentions that you may need to override the setting
in jetty.ini, but if you do this after installing via the module,
the module will override your setting.  This commit takes a stab
at updating the docs to indicate how you'd accomplish this via
the module.

@nfagerlund: I presume you'll want to reword or entirely gut this
verbiage and work your magic on it, but I just wanted to submit this
as a starting point.