Fixup global open_jdk defines
Previously the open_jdk on sles would be set to java-1_6_0-ibm before being set
again to java-1.6.0-openjdk (after exiting the inner if block). This package
won't be available on sles. This commit updates the conditionals to set ibm
java for sles and openjdk 1.6 java for el and fedora < 17.