Remove space between DESTDIR and value
The space between DESTDIR= and $RPM_BUILD_ROOT in the rake call in the rpm spec
file means that DESTDIR gets set to nothing, which later means that puppetdb
and the terminus are installed at top level, which does not work in mock when
building as a non-root user (as it shouldn't). This commit removes the space so
that the DESTDIR is correctly set.