Add a `puppet storeconfigs export` command to dump the database
This command can be used to export a copy of the ActiveRecord
storeconfigs database (only exported resources, and only nodes with
exported resources) which can be imported into PuppetDB.  This is useful
when migrating from old-school storeconfigs to PuppetDB, as it allows
agents to run without losing any of their exported resources for a time.
Once every agent has run once, the database should be up-to-date in any
case.