Acceptance: Remove debugging output from version checking
We were experiencing a weird bug where the acceptance tests were
failing on Debian systems when they were attempting to validate
the version number of the installed package.  We'd added some
debug output to help us narrow down the root cause of the problem.

We've since discovered what the issue was and fixed it, so this
commit simply removes the debugging output.