Write the storeconfigs database dump as a .tar.gz
Previously this was a single file containing many JSON objects, for the
catalogs. Now it's a .tar.gz (or .tar if gzip can't be found [or
directory if tar can't be found]) containing one file per node. The
tarball also includes a metadata file specifying the nodes which are
present, the timestamp of the export, and the version of the "replace
catalog" command which ought to be used to submit the catalogs to
PuppetDB.