(#19239) Work around EOFException upon journal corruption
We discovered a bug in KahaDB that when the journal is corrupted there might be
a stray EOFException that is not caught by the KahaDB store, however upon
broker restart it corrects itself.

This patch will automatically restart the broker when that state is detected,
avoiding the need for an administrator to intervene.

This also adds a section to documentation to explain how to recover from
KahaDB corruption errors manually.

Signed-off-by: Ken Barber <ken@bob.sh>