[packaging] use packaging repo data to determine PE build
Part of the new Release packaging workflow involves sending builds to jenkins
to do our work for us. However, when we build with the packaging repo via
jenkins, we don't pass environment variables up in anyway. Instead, all data is
stored in a build parameters file. With puppetdb, we store whether or not we
are building PE in the build parameters file. This is lost though, because
currently we only check for PE build in the top-level Rakefile by querying the
environment variable PE_BUILD. To enable builds to succeed with the packaging
repo jenkins workflow, we need to also attempt to query the data in the build
parameters which is made available via the @build object via its build_pe
instance variable. This commit updates the Rakefile to query this first if set,
and then fall back to environment variable.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>